+++ python/docs/source/conf.py
@@ -16,8 +16,8 @@ project = 'ADCToolbox'
 copyright = '2026, ADCToolbox Contributors'
 author = 'ADCToolbox Contributors'
 
-version = '0.8.0'
-release = '0.8.0'
+version = '0.8.1'
+release = '0.8.1'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration